Changed unsigned long long int to uint64_t

+++ ft_putmem.c
@@ -6,7 +6,7 @@
 /*   By: mialbert <mialbert@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/06 21:01:19 by mialbert          #+#    #+#             */
-/*   Updated: 2021/12/16 21:52:29 by mialbert         ###   ########.fr       */
+/*   Updated: 2021/12/17 12:04:14 by mialbert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,5 @@ int32_t	ft_putmem(size_t *ptr, const char *hexbase)
 	ptr2 = (size_t)ptr;
 	if (!ptr)
 		return (ft_putstr("0x0"));
-	return (ft_putstr("0x") + ft_hex2((unsigned long long int)ptr, hexbase));
+	return (ft_putstr("0x") + ft_hex2((uint64_t)ptr, hexbase));
 }
